+++ docs/CHANGELOG.md
@@ -9,11 +9,13 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 ## Unreleased
 ### Added
+* Added Named Component. ([#879])
 ### Changed
 * Made get_camera public. ([#878)]
 ### Removed
 ### Fixed
 
+[#879]: https://github.com/amethyst/amethyst/pull/879
 [#878]: https://github.com/amethyst/amethyst/pull/878
 
 ## [0.8.0] - 2018-08